혁주  - 팀기록 완성

+++ baseball/src/main/java/org/bakara/dto/TeamRecordDTO.java
@@ -23,9 +23,9 @@ public class TeamRecordDTO {
 	int ifb; 
 	int so ; 
 	int doubleout ;
-	int avg; 
-	int obp;
-	int sig ; 
-	int ops; 
+	Long avg; 
+	Long obp;
+	Long sig ; 
+	Long ops; 
 	
 }
